Add note that 504 currently returns 200
Stumbled over this issue (https://github.com/Readify/httpstatus/issues/98) when trying to figure out why the /504 page didn't work as expected. I figured it'd be nice to add a note to the official status documentation, since it's been a few months since the 504 noop was enabled and there doesn't seem to be a timeline for resolution.